ShouldThrowDependencyExceptionOnDeleteWhenSqlExceptionOccursAndLogItAsync -> PASS ShouldThrowDependencyExceptionOnDeleteWhenDbExceptionOccursAndLogItAsync -> PASS ShouldThrowServiceExceptionOnDeleteWhenExceptionOccursAndLogItAsync -> PASS ShouldPostStudentAsync -> PASS
